Create Demo.mp4
This is a demo to show our trojan in action. I would have liked to show simulation along with this but I wasn't able to get it to work just yet. In this demo we have the test_seed set to 32'h1 but I forget to scroll to the top to show it.